+++ passkit_ui/lib/src/event_ticket.dart
@@ -169,7 +169,6 @@ class _AuxiliaryRow extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final directionality = Directionality.of(context);
     return Row(
       mainAxisAlignment: MainAxisAlignment.spaceBetween,
       children: auxiliaryRow.map((item) {
@@ -178,16 +177,12 @@ class _AuxiliaryRow extends StatelessWidget {
             Text(
               item.label ?? '',
               style: passTheme.labelTextStyle,
-              textAlign: item.textAlignment?.toFlutterTextAlign(
-                textDirection: directionality,
-              ),
+              textAlign: item.textAlignment.toFlutterTextAlign(),
             ),
             Text(
               item.value.toString(),
               style: passTheme.foregroundTextStyle,
-              textAlign: item.textAlignment?.toFlutterTextAlign(
-                textDirection: directionality,
-              ),
+              textAlign: item.textAlignment.toFlutterTextAlign(),
             ),
           ],
         );