+++ passkit_ui/test/src/widgets/passkit_barcode_test.dart
@@ -14,7 +14,7 @@ void main() {
         PasskitBarcode(
           barcode: pk.Barcode(
             message: '1234567890',
-            format: 'PKBarcodeFormatQR',
+            format: pk.PkPassBarcodeType.qr,
             messageEncoding: 'utf-8',
             altText: altText,
           ),